fix: Resolve type errors in admin pages
This commit addresses two type-related errors that were causing the
build to fail on the admin pages.
- Renames the post status state variable to `postStatus` on the create
page to resolve a naming conflict with the `useSession` hook's `status`
variable.
- Updates the `UKButton` component to correctly accept a `type` prop,
fixing an error when using it as a form submission button.
