+++ app/api/auth/[...nextauth]/route.ts
@@ -1,24 +1,5 @@
 import NextAuth from 'next-auth'
-import GithubProvider from 'next-auth/providers/github'
-
-export const authOptions = {
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID ?? '',
-      clientSecret: process.env.GITHUB_SECRET ?? '',
-    }),
-  ],
-  callbacks: {
-    async signIn({ user, profile }) {
-      // @ts-ignore
-      if (profile?.login === 'uki') {
-        return true
-      } else {
-        return '/unauthorized'
-      }
-    },
-  },
-}
+import { authOptions } from 'app/lib/auth'
 
 const handler = NextAuth(authOptions)
