+++ app/lib/auth.ts
@@ -1,0 +1,21 @@
+import type { NextAuthOptions } from 'next-auth'
+import GithubProvider from 'next-auth/providers/github'
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_ID ?? '',
+      clientSecret: process.env.GITHUB_SECRET ?? '',
+    }),
+  ],
+  callbacks: {
+    async signIn({ user, profile }) {
+      // @ts-ignore
+      if (profile?.login === 'uki') {
+        return true
+      } else {
+        return '/unauthorized'
+      }
+    },
+  },
+}
